Add ability to automatically send yanked text to Clipper
Works on Neovim thanks to the `TextYankPost` autocommand. Can opt out
via `'g:ClipperAuto'`.

This aligns with the set-up I have in recent versions of tmux. Note that
I've explicitly scoped it to explicit yank operations (not deletions or
changes). Given that one tends to yank a lot in Vim, for this to be
usually helpful but with an escape hatch in case the behavior is not
what you want in a specific instance, it's important to pair it with a
tool like Alfred that remembers recent clipboard history, so that you
can go back to a previous entry if you end up inadvertently overwriting
something.