fix: ignore case during string check
What was previously written as "shutdown(2)" in the older version, is
now written as "Shutdown", which breaks the string check and causes the
netcat binary to be invoked without the "-N" option. This fix expands
the check behaviour to ignore the case (\c), which causes the Linux
netcat binary to be properly detected.

Closes: https://github.com/wincent/vim-clipper/issues/4