feat: use OSC 52 sequence when in kitty
This is just a POC at this stage, so I'm initially just committing it on
the "next" branch. As noted by the comment, would probably be better to
do the base64-encoding in-process.

For now though, to keep things minimal, I'm using exactly the same
approach that I am using in my dotfiles in the `clip` helper:

    https://github.com/wincent/wincent/blob/c85890b0dada7ad1611501b7cde3a904aa48764c/aspects/dotfiles/files/.zsh/bin/clip